+++ apidaora/route/annotations_getters.py
@@ -24,7 +24,9 @@ def get_annotations_query_dict(
         if name not in path_args_annotations
         and name not in headers_annotations
         and name != 'return'
+        and name != 'args'
         and name != 'kwargs'
+        and name != 'request'
     }