+++ docs/src/middlewares/middlewares_curl2.bash
@@ -1,1 +1,1 @@
-curl -i localhost:8000/hello-pre-execution -X POST -d '{"name": "Me"}'
+curl -i 'localhost:8000/hello-post-execution?name=Me'