+++ docs/src/middlewares_extra_args/middlewares_extra_args.py
@@ -1,33 +1,23 @@
 from typing import Any
 
-from apidaora import (
-    MiddlewareRequest,
-    Middlewares,
-    Response,
-    appdaora,
-    route,
-    text,
-)
+from apidaora import Middlewares, Request, Response, appdaora, route, text
 
 
-def request_extra_args(request: MiddlewareRequest) -> None:
-    request.kwargs = {'extra': 'You'}
+def request_extra_args(request: Request) -> None:
+    request.ctx['extra'] = 'You'
 
 
-def response_extra_args(
-    request: MiddlewareRequest, response: Response
-) -> None:
-    if request.kwargs and response.kwargs:
+def response_extra_args(request: Request, response: Response) -> None:
+    if request.ctx and response.ctx:
         response.body = response.body.replace(
-            'You', f"{request.kwargs['extra']} and {response.kwargs['name']}"
+            'You', f"{request.ctx['extra']} and {response.ctx['name']}"
         )
 
 
 @route.get(
-    '/middlewares-kwargs',
+    '/middlewares-ctx',
     middlewares=Middlewares(
-        pre_execution=[request_extra_args],
-        post_execution=[response_extra_args],
+        pre_execution=request_extra_args, post_execution=response_extra_args,
     ),
 )
 async def extra_args_controller(name: str, **kwargs: Any) -> Response: