+++ docs/src/middlewares_extra_args/middlewares_extra_args_curl.bash
@@ -1,1 +1,1 @@
-curl -i localhost:8000/middlewares-kwargs?name=Me
+curl -i localhost:8000/middlewares-ctx?name=Me